add julia tutorial to user home directory
The tutorial files are placed in a folder inside `.juliabox` and
a symlink is made from user home folder. Since `.juliabox` folder
does not get backed up, this does not add any load on the backup
process and tutorials get refreshed on each login.

For existing users, the symlink is created only if user does not
already have a folder/file named `tutorial`.