run webserver (nginx) as a docker container
This changes the webserver component of JuliaBox (nginx openresty) to run as a docker container.
It should make it much easier to install and maintain.

It runs as juser:juser inside the container.
It uses the host network stack to avoid network overheads.
Configuration file, static files and certificates are bundled into the docker container.
Log folder can be mounted from the host system if required.

JuliaBox container manager now treats containers with names suffixed with `_jboxsvc` as  server components running as docker containers
and not user sessions.