container managers (engines) as docker containers
This changes the JuliaBox container manager processes - the tornado
application and the JuliaBox daemon to docker containers. They are
now called Engines.

JuliaBox Engines are docker containers that implement container
management.

It is possible to have different container management strategies
for different types of containers, e.g. for interactive sessions,
for applications, for API hosting.

A base engine container provides common functionality. The base
engine runs in privileged mode and provides common functionalities
over a ZMQ based API. Other engines are build on this base, and
communicate with the base over the provided ZMQ interface.