change routing and cookie scheme as follows:
- one cookie (jb_auth)  to indicate user is authenticated (contains user_id)
- one cookie (jb_sess) to indicate session (contains instance_id, container_id)
- one cookie (jp_<portname>) for each exposed port in the container (contains port number)

Each cookie can be verified separately. That makes it possible to dynamically add more ports to a session, which will be useful in bringing up apps in the container.
Ports are named, and can be accessed via named URLs of the form /jci_<portname>.
The webserver routing logic is simpler and does not need hardcoded paths anymore.
TODO: ability to mark a port as websocket