update julia version, packages & package handling
- update Julia 0.4.0-rc2 to 0.4.0
- add Julia 0.5.0-dev
- add message explaining conflict between syetem provided and user installed packages
- provide button fix conflicts by deleting user installed packages (actually by renaming .julia)
- increase Jupyter `KernelRestarter.time_to_dead` from 3 to 6

Failure to start notebooks are most certainly due to package conflicts. The only way to resolve them is to update local packages to acceptable versions. There's a button in the settings page now to rename `~/.julia` to a temporary folder which should resolve this issue for most cases.

Jupyter restarts the kernel when it fails to start. There's a check to prevent endless restarts, but that kicks in only if kernel restarts in less than 3 secs. That's too small in some cases and causes IJulia to be restarted infinitely, taking up 100% of a CPU core and filling up the disk with logs.

Too many of such runaway processes can saturate the host machine, both in terms of CPU and disk IO.

I am now changing it to 6 based on some local tests. But that's still something that can break either ways. The Jupyter KernelRestarter probably needs a better mechanism.

fixes #306