add user_profiles table, auto fill by auth modules
This adds a new table to store user profile that stores a few
attributes to use while communicating with users.

Authentication modules fill attributes if they are available to them as part of the login response.
A record is maintained to indicate the source of each attribute that was populated automatically.

If users are allowed to explicitly set their profile in the future, authentication modules will not
overwrite the values set by users.